# docs/swagger.yaml
definitions:
...
  github_com_imperatorofdwelling_Full-backend_internal_domain_models_stays.Stay:
    properties:
      created_at:
        type: string
      entrance:
        type: string
      floor:
        type: string
      guests:
        type: integer
      house:
        type: string
      id:
        type: string
      is_smoking_prohibited:
        type: boolean
      location_id:
        type: string
      name:
        type: string
      number_of_bathrooms:
        type: integer
      number_of_bedrooms:
        type: integer
      number_of_beds:
        type: integer
      price:
        type: number
      rating:
        type: number
      room:
        type: string
      square:
        type: number
      street:
        type: string
      type:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  github_com_imperatorofdwelling_Full-backend_internal_domain_models_stays.StayEntity:
    properties:
      entrance:
        type: string
      floor:
        type: string
      guests:
        type: integer
      house:
        type: string
      is_smoking_prohibited:
        default: false
        type: boolean
      location_id:
        type: string
      name:
        type: string
      number_of_bathrooms:
        type: integer
      number_of_bedrooms:
        type: integer
      number_of_beds:
        type: integer
      price:
        type: number
      room:
        type: string
      square:
        type: number
      street:
        type: string
      type:
        type: string
      user_id:
        type: string
    required:
    - guests
    - house
    - location_id
    - name
    - number_of_bathrooms
    - number_of_bedrooms
    - number_of_beds
    - price
    - square
    - street
    - type
    - user_id
    type: object
...
  github_com_imperatorofdwelling_Full-backend_internal_domain_models_staysreports.StayReport:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      report_attach:
        type: string
      stay_id:
        type: string
      title:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  github_com_imperatorofdwelling_Full-backend_internal_domain_models_staysreports.StaysReportEntity:
    properties:
      description:
        type: string
      reportID:
        type: string
      stayName:
        type: string
      title:
        type: string
      userName:
        type: string
    type: object
...
  github_com_imperatorofdwelling_Full-backend_internal_domain_models_usersreports.UsersReportEntity:
    properties:
      description:
        type: string
      owner_name:
        type: string
      report_id:
        type: string
      title:
        type: string
      user_name:
        type: string
    type: object
  github_com_imperatorofdwelling_Full-backend_internal_utils_response.ResponseError:
    properties:
      error:
        type: string
    type: object
...
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
paths:
...
  /report:
    get:
      description: Retrieves all stay reports for the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/github_com_imperatorofdwelling_Full-backend_internal_domain_models_staysreports.StaysReportEntity'
            type: array
        "401":
          description: '{"error": "user not logged in"}'
          schema:
            $ref: '#/definitions/github_com_imperatorofdwelling_Full-backend_internal_utils_response.ResponseError'
        "500":
          description: '{"error": "message"}'
          schema:
            $ref: '#/definitions/github_com_imperatorofdwelling_Full-backend_internal_utils_response.ResponseError'
      summary: Get all stay reports
      tags:
      - stays-reports
  /report/{reportId}:
    delete:
      description: Deletes a specific stay report by report ID
      parameters:
      - description: Report ID
        in: path
        name: reportId
        required: true
        type: string
      responses:
        "200":
          description: '{"message": "Stay report was deleted"}'
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: '{"error": "user not logged in"}'
          schema:
            $ref: '#/definitions/github_com_imperatorofdwelling_Full-backend_internal_utils_response.ResponseError'
        "500":
          description: '{"error": "message"}'
          schema:
            $ref: '#/definitions/github_com_imperatorofdwelling_Full-backend_internal_utils_response.ResponseError'
      summary: Delete a stay report
      tags:
      - stays-reports
    patch:
      consumes:
      - multipart/form-data
      description: Updates specific fields of a stay report, such as title, description,
        or image
      parameters:
      - description: Report ID
        in: path
        name: reportId
        required: true
        type: string
      - description: Updated title
        in: formData
        name: title
        type: string
      - description: Updated description
        in: formData
        name: description
        type: string
      - description: Image file (JPEG or PNG)
        in: formData
        name: image
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Updated stays report object
          schema:
            $ref: '#/definitions/github_com_imperatorofdwelling_Full-backend_internal_domain_models_staysreports.StaysReportEntity'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_imperatorofdwelling_Full-backend_internal_utils_response.ResponseError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_imperatorofdwelling_Full-backend_internal_utils_response.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_imperatorofdwelling_Full-backend_internal_utils_response.ResponseError'
      summary: Partially update a stay report
      tags:
      - stays-reports
  /report/create/{stayId}:
    post:
      consumes:
      - multipart/form-data
      description: Creates a report for a specific stay, including an optional image
        and required details like title and description.
      parameters:
      - description: ID of the stay being reported
        in: path
        name: stayId
        required: true
        type: string
      - description: Title of the report
        in: formData
        name: title
        required: true
        type: string
      - description: Description of the report
        in: formData
        name: description
        required: true
        type: string
      - description: Optional image file (JPEG or PNG)
        in: formData
        name: image
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Confirmation message
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Error message for invalid input or unsupported image type
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Error message for unauthorized access
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Error message for internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Create a new stay report
      tags:
      - StaysReports
...
  /stays:
    get:
      consumes:
      - application/json
      description: Get all stays
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            items:
              $ref: '#/definitions/github_com_imperatorofdwelling_Full-backend_internal_domain_models_stays.Stay'
            type: array
        "400":
          description: Error
          schema:
            $ref: '#/definitions/github_com_imperatorofdwelling_Full-backend_internal_utils_response.ResponseError'
        default:
          description: Error
          schema:
            $ref: '#/definitions/github_com_imperatorofdwelling_Full-backend_internal_utils_response.ResponseError'
      summary: Get all stays
      tags:
      - stays
  /stays-reports:
    get:
      description: Fetches a specific stay report associated with the logged-in user.
      produces:
      - application/json
      responses:
        "200":
          description: Retrieved stay report object
          schema:
            $ref: '#/definitions/github_com_imperatorofdwelling_Full-backend_internal_domain_models_staysreports.StayReport'
        "401":
          description: '{"error": "user not logged in"}'
          schema:
            $ref: '#/definitions/github_com_imperatorofdwelling_Full-backend_internal_utils_response.ResponseError'
        "404":
          description: '{"error": "report not found"}'
          schema:
            $ref: '#/definitions/github_com_imperatorofdwelling_Full-backend_internal_utils_response.ResponseError'
        "500":
          description: '{"error": "could not fetch report"}'
          schema:
            $ref: '#/definitions/github_com_imperatorofdwelling_Full-backend_internal_utils_response.ResponseError'
      security:
      - ApiKeyAuth: []
      summary: Retrieve stay report by user ID
      tags:
      - stays-reports
  /stays/{stayId}:
    delete:
      consumes:
      - application/json
      description: Delete stay by id
      parameters:
      - description: stay id
        in: path
        name: stayId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: no content
          schema:
            type: string
        "400":
          description: Error
          schema:
            $ref: '#/definitions/github_com_imperatorofdwelling_Full-backend_internal_utils_response.ResponseError'
        default:
          description: Error
          schema:
            $ref: '#/definitions/github_com_imperatorofdwelling_Full-backend_internal_utils_response.ResponseError'
      summary: Delete Stay
      tags:
      - stays
    get:
      consumes:
      - application/json
      description: get stay
      parameters:
      - description: stay id
        in: path
        name: stayId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/github_com_imperatorofdwelling_Full-backend_internal_domain_models_stays.Stay'
        "400":
          description: Error
          schema:
            $ref: '#/definitions/github_com_imperatorofdwelling_Full-backend_internal_utils_response.ResponseError'
        default:
          description: Error
          schema:
            $ref: '#/definitions/github_com_imperatorofdwelling_Full-backend_internal_utils_response.ResponseError'
      summary: Get Stay by id
      tags:
      - stays
    put:
      consumes:
      - application/json
      description: Update stay by id
      parameters:
      - description: request stay data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_imperatorofdwelling_Full-backend_internal_domain_models_stays.StayEntity'
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/github_com_imperatorofdwelling_Full-backend_internal_domain_models_stays.Stay'
        "400":
          description: Error
          schema:
            $ref: '#/definitions/github_com_imperatorofdwelling_Full-backend_internal_utils_response.ResponseError'
        default:
          description: Error
          schema:
            $ref: '#/definitions/github_com_imperatorofdwelling_Full-backend_internal_utils_response.ResponseError'
      summary: Update Stay
      tags:
      - stays
...
  /user/report/:
    get:
      description: Retrieves all reports created by a user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get All User Reports
      tags:
      - UsersReports
  /user/report/{reportId}:
    delete:
      description: Deletes a specified report created by a user
      parameters:
      - description: ID of the report to delete
        in: path
        name: reportId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete User Report
      tags:
      - UsersReports
    patch:
      consumes:
      - multipart/form-data
      description: Updates specific fields (title, description, or image) of an existing
        user report
      parameters:
      - description: ID of the report to update
        in: path
        name: reportId
        required: true
        type: string
      - description: New title for the report
        in: formData
        name: title
        type: string
      - description: New description for the report
        in: formData
        name: description
        type: string
      - description: Optional new image file (JPEG or PNG)
        in: formData
        name: image
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Updated user report object
          schema:
            $ref: '#/definitions/github_com_imperatorofdwelling_Full-backend_internal_domain_models_usersreports.UsersReportEntity'
        "400":
          description: Invalid request
          schema:
            $ref: '#/definitions/github_com_imperatorofdwelling_Full-backend_internal_utils_response.ResponseError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_imperatorofdwelling_Full-backend_internal_utils_response.ResponseError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_imperatorofdwelling_Full-backend_internal_utils_response.ResponseError'
      summary: Update User Report
      tags:
      - UsersReports
  /user/report/{toBlameId}:
    post:
      consumes:
      - multipart/form-data
      description: Creates a user report with an optional image and necessary details
        such as title and description.
      parameters:
      - description: ID of the user being reported
        in: path
        name: toBlameId
        required: true
        type: string
      - description: Title of the report
        in: formData
        name: title
        required: true
        type: string
      - description: Description of the report
        in: formData
        name: description
        required: true
        type: string
      - description: Optional image file (JPEG or PNG)
        in: formData
        name: image
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Message confirming successful creation
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid input, missing fields, or unsupported image type
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized, user not logged in
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new user report
      tags:
      - User Reports
  /users/report/{reportId}:
    get:
      consumes:
      - application/json
      description: Retrieve detailed information about a specific user's report by
        its unique ID.
      parameters:
      - description: Report ID
        in: path
        name: reportId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successful response containing user report data
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 'Unauthorized: User not logged in'
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get a user's report by ID
      tags:
      - UsersReports
swagger: "2.0"